Added more gRPC requests

+++ src/DistributedFileSystem.MasterNode/Services/MongoDbService.cs
@@ -136,7 +136,7 @@ namespace DistributedFileSystem.MasterNode.Services
         }
 
         // This uses worker's resources and status to help choose a worker that can handle the request with the least amount of latency possible
-        public async Task<string?> GetOptimalWorker(int chunkSize)
+        public async Task<string?> GetOptimalWorker(long chunkSize)
         {
             _logger.LogInformation("Fetching most optimal worker with status 'waiting' and sufficient disk space.");
